Fixed a problem with double starting after a file change. Added a windows load feature to the Load Function, if you call it with a null filename it will now open a Browse window. Improved the method of changing files, can now change files when paused and stopped. Fixed a problem with the property page updating while the pidparser was restarting.
git-svn-id: http://svn.nate.deepcreek.org.au/svn/TSFileSource/trunk@71 89fdcf3f-49f1-7541-bbf8-3596d8354d9d

+++ src/MultiFileWriter.cpp
@@ -317,9 +317,13 @@ HRESULT MultiFileWriter::ReuseTSFile()
 		return hr;
 	}
 
-	TCHAR sz[MAX_PATH];
-	sprintf(sz, "%S", pFilename);
-	DeleteFile(sz);
+	// Check if file is being read by something.
+	if (IsFileLocked(pFilename) != TRUE)
+	{
+		TCHAR sz[MAX_PATH];
+		sprintf(sz, "%S", pFilename);
+		DeleteFile(sz);
+	}
 		
 	if FAILED(hr = m_pCurrentTSFile->OpenFile())
 	{
